@@ -12,15 +12,17 @@
     "devDependencies": {
         "axios": "^0.19",
         "bootstrap": "^4.0.0",
-        "cross-env": "^7.0",
+        "cross-env": "^7.0.3",
         "jquery": "^3.2",
-        "laravel-mix": "^5.0.1",
+        "laravel-mix": "^5.0.9",
         "lodash": "^4.17.19",
         "popper.js": "^1.12",
         "resolve-url-loader": "^2.3.1",
         "sass": "^1.20.1",
         "sass-loader": "^8.0.0",
-        "vue": "^2.5.17",
+        "vue": "^2.6.12",
+        "vue-router": "^3.4.9",
         "vue-template-compiler": "^2.6.10"
-    }
+    },
+    "dependencies": {}
 }
