@@ -1,5 +1,6 @@
 <?php 
 return['mypage_informations'=>[['user.pay.index','ポイントを追加する'],['user.instruction','指示を追加する'],
-        ['user.field','畑を追加する'],['user.plant','植物を追加する'],['user.myfield','マイ畑を見る'],
-        ['user.daily.index', '日記を見る'],['community.view','コミュニティーチャット'],['user_request.create','要望をする'],['chat.create_chatroom','チャットを送る']],
-        'admin_user'=>['realFarmerAdmin@email.com','realFarmerAdmin'],];
+        ['user.field','畑を追加する'],['user.plant','植物を追加する'],
+        ['user.daily.index', '日記を見る'],['community.view','コミュニティーチャット'],['user_request.create','要望をする'],['chat.create_chatroom','チャットを送る'],['user.myfield','自分の畑の情報を見る']],
+        'admin_user'=>['realFarmerAdmin@email.com','realFarmerAdmin'],
+        'mypage_image'=>['/img/point.jpg','/img/instruction.jpg','/img/field.jpg','/img/plant.jpg','/img/plant.jpg','/img/plant.jpg','/img/plant.jpg','/img/plant.jpg','/img/plant.jpg','/img/plant.jpg']];
