@@ -9,7 +9,7 @@
     <form action="{{ route('user.menu.field') }}" method="POST">
          {{ csrf_field() }}
     <p>畑を追加する（複数回答可）: 
-   
+    @if(count($fields)>0)
     @foreach($fields as $index=>$field)
         @if($field->used==true)
             <div class="index-container shadow">
@@ -19,6 +19,8 @@
              </div>
         @endif
     @endforeach
+    @endif 
+
    
     </p>
     <div class="skyblue panel"><button class="skyblue menu-button shadow" type="submit" value="送信">送信</button></div>
